fix login with google

+++ src/pages/settings/users/MemberTab.jsx
@@ -18,28 +18,58 @@ import { fetchAllGroups } from "../../../services/group";
 import PasswordInput from "../../../components/PasswordInput";
 import { resetPasswordById } from "../../../services/auth";
 import useAuth from "../../../hooks/useAuth";
+import { queryClient } from "../../../App";
 
 function MemberTab() {
   const { auth } = useAuth();
   const permission = auth?.permission || [];
   const dialogRef = useRef();
   const notiDialogRef = useRef();
+  const formRef = useRef();
   const [isPasswordChanged, setIsPasswordChanged] = useState(false);
+  const [searchState, setSearchState] = useState({
+    state: "1",
+    textSearch: "",
+  });
   const [dialogState, setDialogState] = useState({
     data: null,
     isEditable: true,
   });
-  const [listState, setListState] = useState([]);
 
   const membersQuery = useQuery({
     queryKey: ["members"],
-    queryFn: fetchAllUsers,
+    queryFn: async () => {
+      const data = await fetchAllUsers();
+
+      const searchData = data.filter(
+        (item) =>
+          item?.fullName.toLowerCase().includes(searchState.textSearch) &&
+          checkSearchState({ state: searchState.state, user: item })
+      );
+
+      return searchData;
+    },
   });
+
+  function checkSearchState({ state, user }) {
+    if (state === "1") {
+      return user.isActive === 1 ? true : false;
+    } else if (state === "2") {
+      return user.isActive === 0 ? true : false;
+    } else {
+      return true;
+    }
+  }
+
   const members = membersQuery.data;
 
   const groupQuery = useQuery({
     queryKey: ["group"],
-    queryFn: fetchAllGroups,
+    queryFn: async () => {
+      const data = await fetchAllGroups();
+      const activeGroup = data.filter((item) => item.isActive === 1);
+      return activeGroup;
+    },
   });
 
   const groups = groupQuery.data;
@@ -55,16 +85,10 @@ function MemberTab() {
   }
 
   useEffect(() => {
-    setListState(() => members);
-  }, [members]);
-
-  function searchHandler(event) {
-    const textSearch = event.target.value.toLowerCase().trim();
-    const result = members.filter((member) =>
-      member.fullName.toLowerCase().includes(textSearch)
-    );
-    setListState(() => result);
-  }
+    queryClient.invalidateQueries({
+      queryKey: ["members"],
+    });
+  }, [searchState]);
 
   async function editUserHandler({ id, action }) {
     setIsPasswordChanged(() => false);
@@ -76,14 +100,32 @@ function MemberTab() {
     await dialogRef.current.edit({ id, action });
   }
 
-  async function deleteUserHandler(id) {
-    notiDialogRef.current.setDialogData({
-      action: DialogAction.DELETE,
-      dispatchFn: () => deleteUserById({ id }),
+  async function deActivateUserHandler({ id, isActive }) {
+    let adminCount = 0;
+    members.map((member) => {
+      if (member.id === 1) {
+        adminCount++;
+      }
     });
-    notiDialogRef.current.showDialogWarning({
-      message: "Xác nhận xóa thành viên?",
-    });
+
+    if (adminCount <= 1) {
+      notiDialogRef.current.toastError({
+        message: "Không thể xóa admin duy nhất",
+      });
+    } else {
+      notiDialogRef.current.setDialogData({
+        action: DialogAction.DELETE,
+        dispatchFn: () => deleteUserById({ id }),
+      });
+      if (isActive === 1)
+        notiDialogRef.current.showDialogWarning({
+          message: "Xác nhận lưu trữ thành viên?",
+        });
+      else
+        notiDialogRef.current.showDialogWarning({
+          message: "Xác nhận kích hoạt thành viên?",
+        });
+    }
   }
 
   function onSubmitPassword({ data, addMutate }) {
@@ -222,6 +264,17 @@ function MemberTab() {
     </>
   );
 
+  function changeFormHandler() {
+    const formData = new FormData(formRef.current);
+    const searchData = Object.fromEntries(formData);
+    const textSearch = searchData.name.trim();
+    const state = searchData.state;
+    setSearchState({
+      textSearch: textSearch,
+      state: state,
+    });
+  }
+
   return (
     <div className="h-100 w-100">
       <NotificationDialog ref={notiDialogRef} keyQuery={["members"]} />
@@ -232,32 +285,46 @@ function MemberTab() {
               <label>Quản lý thành viên</label>
             </div>
             <div className="row gap-3">
-              <div className="col input-group flex-nowrap">
-                <span
-                  className="input-group-text"
-                  id="addon-wrapping"
-                  style={{ backgroundColor: "white" }}
-                >
-                  <svg
-                    xmlns="http://www.w3.org/2000/svg"
-                    width="16"
-                    height="16"
-                    fill="currentColor"
-                    className="bi bi-search"
-                    viewBox="0 0 16 16"
-                  >
-                    <path d="M11.742 10.344a6.5 6.5 0 1 0-1.397 1.398h-.001q.044.06.098.115l3.85 3.85a1 1 0 0 0 1.415-1.414l-3.85-3.85a1 1 0 0 0-.115-.1zM12 6.5a5.5 5.5 0 1 1-11 0 5.5 5.5 0 0 1 11 0" />
-                  </svg>
-                </span>
-                <input
-                  name="name"
-                  type="search"
-                  className="form-control"
-                  placeholder="Họ và tên"
-                  aria-describedby="addon-wrapping"
-                  onInput={searchHandler}
-                />
-              </div>
+              <form className="col" onChange={changeFormHandler} ref={formRef}>
+                <div className="row gap-3" style={{ width: "fit-content" }}>
+                  <div className="col input-group flex-nowrap">
+                    <span
+                      className="input-group-text"
+                      id="addon-wrapping"
+                      style={{ backgroundColor: "white" }}
+                    >
+                      <svg
+                        xmlns="http://www.w3.org/2000/svg"
+                        width="16"
+                        height="16"
+                        fill="currentColor"
+                        className="bi bi-search"
+                        viewBox="0 0 16 16"
+                      >
+                        <path d="M11.742 10.344a6.5 6.5 0 1 0-1.397 1.398h-.001q.044.06.098.115l3.85 3.85a1 1 0 0 0 1.415-1.414l-3.85-3.85a1 1 0 0 0-.115-.1zM12 6.5a5.5 5.5 0 1 1-11 0 5.5 5.5 0 0 1 11 0" />
+                      </svg>
+                    </span>
+                    <input
+                      name="name"
+                      type="search"
+                      className="form-control"
+                      placeholder="Họ và tên"
+                      aria-describedby="addon-wrapping"
+                    />
+                  </div>
+                  <div className="col input-group flex-nowrap">
+                    <select
+                      className="form-select"
+                      name="state"
+                      defaultValue={1}
+                    >
+                      <option value="1">Hoạt động</option>
+                      <option value="2">Nghỉ việc</option>
+                      <option value="3">Tất cả</option>
+                    </select>
+                  </div>
+                </div>
+              </form>
               <div style={{ width: "fit-content" }}>
                 <MainDialog
                   ref={dialogRef}
@@ -282,7 +349,7 @@ function MemberTab() {
               <div className="text-start" style={{ width: "5%" }}>
                 STT
               </div>
-              <div className="text-start" style={{ width: "19%" }}>
+              <div className="text-start" style={{ width: "14%" }}>
                 Họ và tên
               </div>
               <div className="text-start" style={{ width: "20%" }}>
@@ -291,17 +358,20 @@ function MemberTab() {
               <div className="text-start" style={{ width: "25%" }}>
                 Email
               </div>
-              <div className="text-start" style={{ width: "20%" }}>
+              <div className="text-start" style={{ width: "15%" }}>
                 Vai trò
               </div>
               <div className="text-start" style={{ width: "10%" }}>
+                Trạng thái
+              </div>
+              <div className="text-end" style={{ width: "10%" }}>
                 Thao tác
               </div>
-              <div className="text-start" style={{ width: "1%" }}></div>
+              <div className="text-end" style={{ width: "1%" }}></div>
             </TableHeader>
             <TableBody>
-              {listState &&
-                listState.map((user, index) => {
+              {members &&
+                members.map((user, index) => {
                   return (
                     <li
                       className=" dropdown-center list-group-item list-group-item-primary list-group-item-action w-100 d-flex flex-row"
@@ -316,7 +386,7 @@ function MemberTab() {
                       </div>
                       <div
                         className="text-start"
-                        style={{ width: "20%" }}
+                        style={{ width: "15%" }}
                         aria-expanded="false"
                       >
                         {user.fullName}
@@ -337,7 +407,7 @@ function MemberTab() {
                       </div>
                       <div
                         className="text-start"
-                        style={{ width: "20%" }}
+                        style={{ width: "15%" }}
                         aria-expanded="false"
                       >
                         {user.userGroup.groupName}
@@ -347,52 +417,66 @@ function MemberTab() {
                         style={{ width: "10%" }}
                         aria-expanded="false"
                       >
-                        {permission?.includes("UUser") && (
-                          <span
-                            className="p-2"
-                            onClick={() =>
-                              editUserHandler({ id: user.id, action: "edit" })
-                            }
-                          >
-                            <svg
-                              xmlns="http://www.w3.org/2000/svg"
-                              width="16"
-                              height="16"
-                              fill="#1B59F8"
-                              className="bi bi-pencil-square"
-                              viewBox="0 0 16 16"
+                        {user.isActive === 1 ? "Hoạt động" : "Nghỉ việc"}
+                      </div>
+                      <div
+                        className="text-end"
+                        style={{ width: "10%" }}
+                        aria-expanded="false"
+                      >
+                        {permission?.includes("UUser") &&
+                          user.isActive === 1 && (
+                            <span
+                              className="p-2"
+                              onClick={() =>
+                                editUserHandler({ id: user.id, action: "edit" })
+                              }
                             >
-                              <path d="M15.502 1.94a.5.5 0 0 1 0 .706L14.459 3.69l-2-2L13.502.646a.5.5 0 0 1 .707 0l1.293 1.293zm-1.75 2.456-2-2L4.939 9.21a.5.5 0 0 0-.121.196l-.805 2.414a.25.25 0 0 0 .316.316l2.414-.805a.5.5 0 0 0 .196-.12l6.813-6.814z" />
-                              <path d="M1 13.5A1.5 1.5 0 0 0 2.5 15h11a1.5 1.5 0 0 0 1.5-1.5v-6a.5.5 0 0 0-1 0v6a.5.5 0 0 1-.5.5h-11a.5.5 0 0 1-.5-.5v-11a.5.5 0 0 1 .5-.5H9a.5.5 0 0 0 0-1H2.5A1.5 1.5 0 0 0 1 2.5z" />
-                            </svg>
-                          </span>
-                        )}
-                        {permission?.includes("UUser") && (
+                              <svg
+                                xmlns="http://www.w3.org/2000/svg"
+                                width="16"
+                                height="16"
+                                fill="#1B59F8"
+                                className="bi bi-pencil-square"
+                                viewBox="0 0 16 16"
+                              >
+                                <path d="M15.502 1.94a.5.5 0 0 1 0 .706L14.459 3.69l-2-2L13.502.646a.5.5 0 0 1 .707 0l1.293 1.293zm-1.75 2.456-2-2L4.939 9.21a.5.5 0 0 0-.121.196l-.805 2.414a.25.25 0 0 0 .316.316l2.414-.805a.5.5 0 0 0 .196-.12l6.813-6.814z" />
+                                <path d="M1 13.5A1.5 1.5 0 0 0 2.5 15h11a1.5 1.5 0 0 0 1.5-1.5v-6a.5.5 0 0 0-1 0v6a.5.5 0 0 1-.5.5h-11a.5.5 0 0 1-.5-.5v-11a.5.5 0 0 1 .5-.5H9a.5.5 0 0 0 0-1H2.5A1.5 1.5 0 0 0 1 2.5z" />
+                              </svg>
+                            </span>
+                          )}
+                        {permission?.includes("UUser") &&
+                          user.isActive === 1 && (
+                            <span
+                              className="p-2"
+                              onClick={() =>
+                                changePasswordHandler({
+                                  id: user.id,
+                                  action: "edit",
+                                })
+                              }
+                            >
+                              <svg
+                                xmlns="http://www.w3.org/2000/svg"
+                                width="16"
+                                height="16"
+                                fill="#1B59F8"
+                                className="bi bi-lock-fill"
+                                viewBox="0 0 16 16"
+                              >
+                                <path d="M8 1a2 2 0 0 1 2 2v4H6V3a2 2 0 0 1 2-2m3 6V3a3 3 0 0 0-6 0v4a2 2 0 0 0-2 2v5a2 2 0 0 0 2 2h6a2 2 0 0 0 2-2V9a2 2 0 0 0-2-2" />
+                              </svg>
+                            </span>
+                          )}
+                        {permission?.includes("DUser") && (
                           <span
                             className="p-2"
                             onClick={() =>
-                              changePasswordHandler({
+                              deActivateUserHandler({
                                 id: user.id,
-                                action: "edit",
+                                isActive: user.isActive,
                               })
                             }
-                          >
-                            <svg
-                              xmlns="http://www.w3.org/2000/svg"
-                              width="16"
-                              height="16"
-                              fill="#1B59F8"
-                              className="bi bi-lock-fill"
-                              viewBox="0 0 16 16"
-                            >
-                              <path d="M8 1a2 2 0 0 1 2 2v4H6V3a2 2 0 0 1 2-2m3 6V3a3 3 0 0 0-6 0v4a2 2 0 0 0-2 2v5a2 2 0 0 0 2 2h6a2 2 0 0 0 2-2V9a2 2 0 0 0-2-2" />
-                            </svg>
-                          </span>
-                        )}
-                        {/* {permission?.includes("DUser") && (
-                          <span
-                            className="p-2"
-                            onClick={() => deleteUserHandler(user.id)}
                           >
                             <svg
                               xmlns="http://www.w3.org/2000/svg"
@@ -405,7 +489,7 @@ function MemberTab() {
                               <path d="M12.643 15C13.979 15 15 13.845 15 12.5V5H1v7.5C1 13.845 2.021 15 3.357 15zM5.5 7h5a.5.5 0 0 1 0 1h-5a.5.5 0 0 1 0-1M.8 1a.8.8 0 0 0-.8.8V3a.8.8 0 0 0 .8.8h14.4A.8.8 0 0 0 16 3V1.8a.8.8 0 0 0-.8-.8z" />
                             </svg>
                           </span>
-                        )} */}
+                        )}
                       </div>
                     </li>
                   );
